Move several types to common package
