--- alpaca/broker/constants.py
@@ -1,2 +1,0 @@
-#
-ACCOUNT_ACTIVITIES_DEFAULT_PAGE_SIZE = 100
