+++ alpaca/common/constants.py
@@ -1,1 +1,3 @@
 DATA_V2_MAX_LIMIT = 10000  # max items per api call
+
+ACCOUNT_ACTIVITIES_DEFAULT_PAGE_SIZE = 100
