+++ alpaca/common/models/accounts.py
@@ -1,0 +1,99 @@
+from datetime import datetime, date
+from typing import Optional
+from uuid import UUID
+
+from .models import ValidateBaseModel as BaseModel
+
+from ..enums import (
+    ActivityType,
+    NonTradeActivityStatus,
+    OrderStatus,
+    OrderSide,
+    TradeActivityType,
+)
+
+
+class BaseActivity(BaseModel):
+    """
+    Represents Base information for an event/activity for a specific Account.
+
+    You most likely will want an instance of one of the child classes TradeActivity and NonTradeActivity
+
+    Attributes:
+        id (str): Unique ID of this Activity. Note that IDs for Activity instances are formatted like
+          `20220203000000000::045b3b8d-c566-4bef-b741-2bf598dd6ae7` the first part before the `::` is a date string
+          while the part after is a UUID
+        account_id (UUID): id of the Account this activity relates too
+        activity_type (ActivityType): What specific kind of Activity this was
+    """
+
+    id: str
+    account_id: UUID
+    activity_type: ActivityType
+
+    def __init__(self, *args, **data):
+        if "account_id" in data and type(data["account_id"]) == str:
+            data["account_id"] = UUID(data["account_id"])
+
+        super().__init__(*args, **data)
+
+
+class NonTradeActivity(BaseActivity):
+    """
+    A NonTradeActivity represents an Activity that happened for an account that doesn't have to do with orders or trades.
+
+    Attributes:
+        date (date): The date on which the activity occurred or on which the transaction associated with the
+          activity settled.
+        net_amount (float): The net amount of money (positive or negative) associated with the activity.
+        description (str): Extra description of the NTA if needed. Can be empty string ""
+        status (NonTradeActivityStatus): Status of the activity. Not present for all activity types.
+        symbol (Optional[str]): The symbol of the security involved with the activity. Not present for all activity
+          types.
+        qty (Optional[float]): For dividend activities, the number of shares that contributed to the payment. Not
+          present for other activity types.
+        price (Optional[float]) Not present for all activity types.
+        per_share_amount (Optional[float]): For dividend activities, the average amount paid per share. Not present for
+          other activity types.
+    """
+
+    date: date
+    net_amount: float
+    description: str
+    status: Optional[NonTradeActivityStatus] = None
+    symbol: Optional[str] = None
+    qty: Optional[float] = None
+    price: Optional[float] = None
+    per_share_amount: Optional[float] = None
+
+
+class TradeActivity(BaseActivity):
+    """
+    Represents information for TradeActivities. TradeActivities are Activities that pertain to trades that happened for
+    an account. IE Fills or partial fills for orders.
+
+    Attributes:
+        transaction_time (datetime): The time and date of when this trade was processed
+        type (TradeActivityType): What kind of trade this TradeActivity represents. See TradeActivityType for more
+          details
+        price (float): The per-share price that the trade was executed at.
+        qty (float): The number of shares involved in the trade execution.
+        side (OrderSide): What side the trade this TradeActivity represents was on. See OrderSide for more information
+        symbol (str): The symbol of the asset that was traded
+        leaves_qty (float): For partially filled orders, the quantity of shares that are left to be filled. Will be 0 if
+          order was not a partially filled order
+        order_id (UUID): The ID for the order filled
+        cum_qty (float): The cumulative quantity of shares involved in the execution.
+        order_status (OrderStatus): The status of the order that executed the trade. See OrderStatus for more details
+    """
+
+    transaction_time: datetime
+    type: TradeActivityType
+    price: float
+    qty: float
+    side: OrderSide
+    symbol: str
+    leaves_qty: float
+    order_id: UUID
+    cum_qty: float
+    order_status: OrderStatus
